NEW Save/load info from CSV of IDs
Allow mapping upload file IDs to a non-relationship field on a data object
as a comma separated list of IDs (and loading attachments from the same field)
instead of requiring a relationship to exist on a record.

Makes it easier to incorporate in other form types where a record doesn't
necessarily exist at the time of loading the form, but a value can be
set.